requirements.txt
numpy==1.11.3
olefile==0.43
pbr==1.10.0
Pillow==4.0.0
protobuf==3.1.0.post1
PyYAML==3.12
scipy==0.18.1
six==1.10.0
tensorflow==0.12.1
